refactor(model)!: move `Embed` under `message` module
