+++ cache/in-memory/src/model/message.rs
@@ -4,10 +4,10 @@ use serde::Serialize;
 use twilight_model::{
     application::interaction::InteractionType,
     channel::{
-        embed::Embed,
         message::{
-            sticker::MessageSticker, Component, Message, MessageActivity, MessageApplication,
-            MessageFlags, MessageInteraction, MessageReaction, MessageReference, MessageType,
+            sticker::MessageSticker, Component, Embed, Message, MessageActivity,
+            MessageApplication, MessageFlags, MessageInteraction, MessageReaction,
+            MessageReference, MessageType,
         },
         Attachment, ChannelMention,
     },
