+++ http/src/request/application/interaction/update_response.rs
@@ -12,10 +12,7 @@ use crate::{
 };
 use serde::Serialize;
 use twilight_model::{
-    channel::{
-        embed::Embed,
-        message::{AllowedMentions, Component, Message},
-    },
+    channel::message::{AllowedMentions, Component, Embed, Message},
     http::attachment::Attachment,
     id::{
         marker::{ApplicationMarker, AttachmentMarker},
