+++ model/src/channel/message/embed/footer.rs
