+++ model/src/channel/message/embed/image.rs
