+++ model/src/channel/message/embed/mod.rs
