+++ model/src/channel/message/embed/provider.rs
