+++ model/src/channel/message/embed/thumbnail.rs
