+++ model/src/gateway/payload/incoming/message_update.rs
@@ -1,7 +1,6 @@
 use crate::{
     channel::{
-        embed::Embed,
-        message::{Mention, MessageType},
+        message::{Embed, Mention, MessageType},
         Attachment,
     },
     id::{
