+++ util/src/builder/interaction_response_data.rs
@@ -1,9 +1,6 @@
 use twilight_model::{
     application::command::CommandOptionChoice,
-    channel::{
-        embed::Embed,
-        message::{AllowedMentions, Component, MessageFlags},
-    },
+    channel::message::{AllowedMentions, Component, Embed, MessageFlags},
     http::{attachment::Attachment, interaction::InteractionResponseData},
 };
